move getLengthID to helperFunctions/lengths.ts
